+++ src/frontend/web_application/stories/CHANGELOG.md
@@ -7,6 +7,7 @@
   - DropdownMenu
   - MultidimensionalPi
   - IconLetter simple refactor of ContactIconLetter
+  - Layout - TabList Presenter
 
 ### Changed